requirements.txt
Django==1.9.4
django-autocomplete-light==3.1.6
django-bootstrap3==7.0.1
django-crispy-forms==1.6.0
django-filter==0.13.0
django-forms-bootstrap==3.0.1
django-grappelli==2.8.1
django-mptt==0.8.3
django-uuslug==1.1.7
djangorestframework==3.3.2
elasticsearch==2.3.0
fake-factory==0.5.7
ipaddress==1.0.16
Markdown==2.6.5
MySQL-python==1.2.5
Pillow==3.1.1
python-dateutil==2.5.2
python-slugify==1.2.0
requests==2.10.0
six==1.10.0
Unidecode==0.4.19
urllib3==1.14
